Improve performance and memory usage of editing Shapes layer  (#8006)
# References and relevant issues

Pointed
https://github.com/napari/napari/pull/7995#issuecomment-2955694555

# Description

Instead of storing information about which vertex/triangle belongs to a
given shape, store information about which range of vertices/triangles
belongs to a given shape.

When editing a shape, try to do as much updating in place as possible
instead of splitting and concatenating the array.

Update typing in `shape_list.py`

A benchmark showing that the used solution is effective.

Use `batched_updates` context manager to reduce call of
`_update_displayed`

---------

Co-authored-by: Carol Willing <carolcode@willingconsulting.com>